reduce tokens from env context (closes #31)
Formerly:

* if no selection: describe every data frame in your global environment
* if a selection: describe every variable in your environment that also appears in the selection

This resulted in many more input tokens when users didn't make any selections, especially when users have many data frames in their global env.

Now:

* describe every variable that's mentioned either in the selection or in the inputted text
     * if that variable is a data frame, only information on the first 5 rows and 20 columns is presented

Overall, this will cut down on the overall number of tokens sent off to models and should make the number of tokens similar regardless of whether the user supplies a selection or not.